adding gasPrice for anvil,gas optimisations

+++ foundry-fund-me-f23/script/HelperConfig.s.sol
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.18;
-import {Script,console} from "forge-std/Script.sol";
+
+import {Script, console} from "forge-std/Script.sol";
 import {MockV3Aggregator} from "../test/mocks/MockV3Aggregator.sol";
 
 contract HelperConfig is Script {
@@ -9,7 +10,7 @@ contract HelperConfig is Script {
     NetworkConfig public activeNetworkConfig;
     uint8 public constant DECIMALS = 8;
     int256 public constant INITIAL_PRICE = 2000e8;
-    uint public constant SEPOLIA_CHAIN_ID = 11155111;
+    uint256 public constant SEPOLIA_CHAIN_ID = 11155111;
 
     struct NetworkConfig {
         address priceFeed;
@@ -18,22 +19,21 @@ contract HelperConfig is Script {
     constructor() {
         if (block.chainid == SEPOLIA_CHAIN_ID) {
             activeNetworkConfig = getSepoliaEthConfig();
-        } else  {
+        } else {
             activeNetworkConfig = getOrCreateAnvilEthConfig();
         }
     }
-    function getSepoliaEthConfig() public pure returns(NetworkConfig memory) {
+
+    function getSepoliaEthConfig() public pure returns (NetworkConfig memory) {
         // price feed address for sepolia
-        NetworkConfig memory sepoliaConfig = NetworkConfig({
-            priceFeed: 0x694AA1769357215DE4FAC081bf1f309aDC325306
-        });
+        NetworkConfig memory sepoliaConfig = NetworkConfig({priceFeed: 0x694AA1769357215DE4FAC081bf1f309aDC325306});
         return sepoliaConfig;
     }
 
     function getOrCreateAnvilEthConfig() public returns (NetworkConfig memory) {
         // price feed address for anvil
         console.log("What is the address of the price feed?", activeNetworkConfig.priceFeed);
-        if( activeNetworkConfig.priceFeed != address(0)){
+        if (activeNetworkConfig.priceFeed != address(0)) {
             return activeNetworkConfig;
         }
 
@@ -43,9 +43,7 @@ contract HelperConfig is Script {
         vm.startBroadcast();
         MockV3Aggregator mockPriceFeed = new MockV3Aggregator(DECIMALS, INITIAL_PRICE);
         vm.stopBroadcast();
-        NetworkConfig memory anvilConfig = NetworkConfig({
-            priceFeed: address(mockPriceFeed)
-        });
+        NetworkConfig memory anvilConfig = NetworkConfig({priceFeed: address(mockPriceFeed)});
         return anvilConfig;
     }
 }
